feat: 我的技能页面只显示工作区实际文件夹内容，不再读取工程外目录

+++ mvp/src/main/java/net/ooder/mvp/skill/scene/service/impl/LlmStatsServiceImpl.java
@@ -350,7 +350,7 @@ public class LlmStatsServiceImpl implements LlmStatsService {
         stats.setFailedCalls(logs.stream().filter(l -> "error".equals(l.getStatus())).count());
         stats.setTotalTokens(logs.stream().mapToLong(LlmCallLogDTO::getTotalTokens).sum());
         stats.setTotalCost(logs.stream().mapToDouble(LlmCallLogDTO::getCost).sum());
-        stats.setAvgLatency(logs.stream().mapToLong(LlmCallLogDTO::getLatency).average().orElse(0));
+        stats.setAvgLatency((long) logs.stream().mapToLong(LlmCallLogDTO::getLatency).average().orElse(0));
         stats.setSuccessRate(logs.size() > 0 
             ? (double) logs.stream().filter(l -> "success".equals(l.getStatus())).count() / logs.size() * 100 : 0);
